feature/122 refactor: PostDetailModal -> PostModal
